Move files in lib
